Move examples from sql package
