+++ examples/java/sql/src/main/java/org/apache/beam/examples/SqlTransformExample.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.sql;
+package org.apache.beam.examples;
 
 // beam-playground:
 //   name: SqlTransform
